move processing, init and csv handling functions out of main
