+++ csv_client.py
@@ -1,0 +1,36 @@
+from application_mode import application_mode
+
+import csv
+
+
+def read_from_csv():
+    with open('model/keypoint_classifier/keypoint_classifier_label.csv',
+              encoding='utf-8-sig') as f:
+        keypoint_classifier_labels = csv.reader(f)
+        keypoint_classifier_labels = [
+            row[0] for row in keypoint_classifier_labels
+        ]
+    with open(
+            'model/point_history_classifier/point_history_classifier_label.csv',
+            encoding='utf-8-sig') as f:
+        point_history_classifier_labels = csv.reader(f)
+        point_history_classifier_labels = [
+            row[0] for row in point_history_classifier_labels
+        ]
+    return keypoint_classifier_labels, point_history_classifier_labels
+
+
+def log_to_csv(number, mode, landmark_list, point_history_list):
+    if mode == 0:
+        pass
+    if mode == application_mode['LEARN_KEY_POINTS'] and (0 <= number <= 9):
+        csv_path = 'model/keypoint_classifier/keypoint.csv'
+        with open(csv_path, 'a', newline="") as f:
+            writer = csv.writer(f)
+            writer.writerow([number, *landmark_list])
+    if mode == application_mode['LEARN_POINT_HISTORY'] and (0 <= number <= 9):
+        csv_path = 'model/point_history_classifier/point_history.csv'
+        with open(csv_path, 'a', newline="") as f:
+            writer = csv.writer(f)
+            writer.writerow([number, *point_history_list])
+    return
